Add `@jit` decorators to jax.numpy operators.
By wrapping common operators in `jit`, we get a number of benefits:
* `jit` has a faster, more optimized dispatch path compared to the primitive dispatch path in JAX. It's faster to dispatch a `jit` computation than a single primitive.
* `jit` allows us to cache and reuse logic such as broadcasting and type promotion.

One downside is that we now report an error when large Python integer scalars (e.g. `2**32 - 1`) are passed as arguments to JAX array operators. The workaround to this is to use explicitly typed constants instead of Python scalars.

PiperOrigin-RevId: 388551372

+++ tests/lax_numpy_test.py
@@ -3441,7 +3441,7 @@ class LaxBackedNumpyTests(jtu.JaxTestCase):
 
     # explicit uint64 should work
     if config.x64_enabled:
-      self.assertEqual(val, jnp.array(val, dtype='uint64'))
+      self.assertEqual(np.uint64(val), jnp.array(val, dtype='uint64'))
 
   # TODO(jakevdp): fix list inputs to jnp.array and enable the following test
   # def testArrayFromList(self):
